Added add Fluid Regulator

+++ src/lib/db/sqlite/conn.rs
@@ -1,7 +1,7 @@
 use async_trait::async_trait;
 use log;
-use tokio_rusqlite::Connection;
 use std::path::Path;
+use tokio_rusqlite::Connection;
 
 use crate::db::{DatabaseTransactionsFactory, DbConnection};
 use crate::parsers::settings::{self, SqliteConfigurer};
@@ -14,10 +14,9 @@ pub struct OpenSqliteConnection {
 
 #[async_trait]
 impl DbConnection for OpenSqliteConnection {
-    async fn insert(&self, _stmt: String) -> UdmResult<i64> {
+    async fn insert(&self, _stmt: String) -> UdmResult<i32> {
         todo!()
     }
-
 }
 
 impl OpenSqliteConnection {
